+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.1] - 2026-05-07
+
+### Fixed
+- ZIP update now removes old version files before extracting new archive (pre-cleanup via old manifest)
+- Added `WPF LCD Test.exe` to legacy cleanup list for `screen_checker` product after rename
+
+---
+
 ## [1.0.0] - 2026-01-23
 
 Initial release.